% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlj.r
\name{gasinh}
\alias{gasinh}
\title{Generalized Inverse Hyperbolic Sine Transformation}
\usage{
gasinh(x, xi = 0, lambda = 1)
}
\arguments{
\item{x}{A numeric vector to transform.}

\item{xi}{Numeric, location parameter.}

\item{lambda}{Numeric, scale parameter.}
}
\value{
A numeric vector of transformed values.
}
\description{
Computes the generalized inverse hyperbolic sine transformation.
}
