% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlj.r
\name{inverse.gasinh}
\alias{inverse.gasinh}
\title{Inverse of Generalized Inverse Hyperbolic Sine}
\usage{
inverse.gasinh(y, xi = 0, lambda = 1)
}
\arguments{
\item{y}{A numeric vector to transform back.}

\item{xi}{Numeric, location parameter.}

\item{lambda}{Numeric, scale parameter.}
}
\value{
A numeric vector of back-transformed values.
}
\description{
Computes the inverse of the generalized inverse hyperbolic sine transformation.
}
