% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlj.r
\name{lm.gat}
\alias{lm.gat}
\title{Generalized Asinh Transformation (GAT) Model}
\usage{
lm.gat(formula, data = NULL, iterations = 1, penalty = 1e-09, verbose = FALSE)
}
\arguments{
\item{formula}{An object of class \code{formula}.}

\item{data}{An optional data frame.}

\item{iterations}{Number of iterations for optimization.}

\item{penalty}{Penalty for regularization.}

\item{verbose}{Logical, if TRUE, prints progress.}
}
\value{
A list containing:
\describe{
\item{predict}{A function to predict new data.}
\item{detransform}{A function to reverse the transformation.}
\item{par}{Optimized transformation parameters.}
\item{fit}{Fitted linear model object.}
\item{opt}{Optimization object.}
\item{df.sub}{Subset of data frame used in modeling.}
\item{df.mut}{Transformed data frame.}
\item{z}{Detransformed fitted values.}
}
}
\description{
Fits a linear model using the generalized asinh transformation.
}
