% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlj.r
\name{pjohnson}
\alias{pjohnson}
\title{Cumulative Distribution of Johnson SU}
\usage{
pjohnson(q, gamma = 0, delta = 1, xi = 0, lambda = 1)
}
\arguments{
\item{q}{A numeric vector of quantiles.}

\item{gamma}{Numeric, shape parameter.}

\item{delta}{Numeric, shape parameter.}

\item{xi}{Numeric, location parameter.}

\item{lambda}{Numeric, scale parameter.}
}
\value{
A numeric vector of cumulative probabilities.
}
\description{
Computes the cumulative distribution function of the Johnson SU distribution.
}
