% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlj.r
\name{qqtad}
\alias{qqtad}
\title{Total Absolute Deviation from Theoretical CDF}
\usage{
qqtad(p)
}
\arguments{
\item{p}{A numeric vector representing probabilities.}
}
\value{
A single numeric value representing the total absolute deviation.
}
\description{
Calculates the total absolute deviation from a theoretical cumulative distribution function (CDF).
}
