% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlj.r
\name{rjohnson}
\alias{rjohnson}
\title{Random Deviates from Johnson SU Distribution}
\usage{
rjohnson(n, gamma = 0, delta = 1, xi = 0, lambda = 1)
}
\arguments{
\item{n}{Integer, the number of random values to generate.}

\item{gamma}{Numeric, shape parameter (default = 0).}

\item{delta}{Numeric, shape parameter (default = 1).}

\item{xi}{Numeric, location parameter (default = 0).}

\item{lambda}{Numeric, scale parameter (default = 1).}
}
\value{
A numeric vector of random deviates from the Johnson SU distribution.
}
\description{
Generates random deviates from the Johnson SU distribution.
}
