% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlj.r
\name{sum_log_cosh}
\alias{sum_log_cosh}
\title{Sum of Log-Cosh Loss}
\usage{
sum_log_cosh(m)
}
\arguments{
\item{m}{A linear model object, which should contain residuals.}
}
\value{
A single numeric value representing the sum of the log-cosh of the residuals.
}
\description{
Computes the sum of the log-cosh of residuals, a smooth approximation to absolute error loss.
}
