% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlj.r
\name{znorm}
\alias{znorm}
\title{Z-Score Normalization}
\usage{
znorm(x)
}
\arguments{
\item{x}{A numeric vector to normalize.}
}
\value{
A list containing:
\describe{
\item{normalize}{A function that applies the normalization to a new vector.}
\item{denormalize}{A function that reverses the normalization.}
\item{par}{A numeric vector of parameters (mean and standard deviation).}
\item{fd}{A function to compute the density under Z-normalization.}
\item{fp}{A function to compute cumulative probability under Z-normalization.}
\item{fq}{A function to compute quantiles under Z-normalization.}
\item{fr}{A function to generate random deviates under Z-normalization.}
\item{x}{The original input data.}
\item{z}{The normalized data.}
}
}
\description{
Normalizes data using Z-score normalization.
}
