% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlj.r
\name{lm.gat}
\alias{lm.gat}
\title{Generalized Asinh Transformation Model (GATM)}
\usage{
lm.gat(
  formula,
  data = NULL,
  loss = sum_log_cosh,
  iterations = 1,
  penalty = 1e-09,
  verbose = FALSE
)
}
\arguments{
\item{formula}{An object of class \code{formula} for the model.}

\item{data}{An optional data frame for variable data.}

\item{iterations}{Integer, number of iterations for optimization (default = 1).}

\item{penalty}{Numeric, regularization penalty (default = 1e-9).}

\item{verbose}{Logical, if TRUE, prints optimization progress (default = FALSE).}
}
\value{
A list containing:
\describe{
\item{detransform}{A function to reverse the generalized asinh transformation.}
\item{par}{Optimized transformation parameters for each variable.}
\item{fit}{Fitted linear model object.}
\item{opt}{Optimization result object.}
\item{df.sub}{Subset of data used for modeling.}
\item{df.mut}{Transformed data frame used in modeling.}
\item{z}{Detransformed fitted values of the model.}
}
}
\description{
Fits a linear model using the generalized asinh transformation.
}
