% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlj.r
\name{qjohnson}
\alias{qjohnson}
\title{Inverse Cumulative Distribution of Johnson SU}
\usage{
qjohnson(p, gamma = 0, delta = 1, xi = 0, lambda = 1)
}
\arguments{
\item{p}{A numeric vector of probabilities.}

\item{gamma}{Numeric, shape parameter (default = 0).}

\item{delta}{Numeric, shape parameter (default = 1).}

\item{xi}{Numeric, location parameter (default = 0).}

\item{lambda}{Numeric, scale parameter (default = 1).}
}
\value{
A numeric vector of quantiles for the Johnson SU distribution.
}
\description{
Computes the inverse cumulative distribution (quantile function) of the Johnson SU distribution.
}
