% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlj.r
\name{rjohnson}
\alias{rjohnson}
\title{Random Deviates from Johnson SU Distribution}
\usage{
rjohnson(n, gamma = 0, delta = 1, xi = 0, lambda = 1)
}
\arguments{
\item{n}{Number of random values to generate.}

\item{gamma}{Numeric, shape parameter.}

\item{delta}{Numeric, shape parameter.}

\item{xi}{Numeric, location parameter.}

\item{lambda}{Numeric, scale parameter.}
}
\value{
A numeric vector of random deviates.
}
\description{
Generates random deviates from the Johnson SU distribution.
}
